구성원 Redis적용

+++ src/main/java/com/sparta/nexusteam/employee/service/DepartmentServiceImpl.java
@@ -2,8 +2,14 @@ package com.sparta.nexusteam.employee.service;
 
 import com.sparta.nexusteam.employee.entity.Department;
 import com.sparta.nexusteam.employee.entity.Employee;
+import com.sparta.nexusteam.employee.entity.UserRole;
 import com.sparta.nexusteam.employee.repository.DepartmentRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.CacheEvict;
+import org.springframework.cache.annotation.CachePut;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.cache.annotation.Caching;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,19 +21,33 @@ public class DepartmentServiceImpl implements DepartmentService {
     private final DepartmentRepository departmentRepository;
 
     @Override
+    @Cacheable(value = "departments", key = "#employee.company.id")
     public List<Department> getAllDepartments(Employee employee) {
         return departmentRepository.findAllByCompany(employee.getCompany());
     }
 
     @Override
+    @Caching(
+            put = @CachePut(value = "department", key = "#result.id"),
+            evict = @CacheEvict(value = "departments", key = "#employee.company.id", allEntries = true)
+    )
     public Department createDepartment(Department department, Employee employee) {
         if (departmentRepository.existsByNameAndCompany(department.getName(), employee.getCompany())) {
             throw new IllegalArgumentException("이미 존재하는 이름입니다");
         }
+
+        if(!UserRole.MANAGER.equals(employee.getRole())){
+            throw new AccessDeniedException("권한이 없습니다");
+        }
+
         return departmentRepository.save(department);
     }
 
     @Override
+    @Caching(
+            put = @CachePut(value = "department", key = "#id"),
+            evict = @CacheEvict(value = "departments", key = "#employee.company.id", allEntries = true)
+    )
     public Department updateDepartment(Long id, Department departmentDetails, Employee employee) {
         Department department = departmentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("부서를 찾을 수 없습니다 " + id));
@@ -35,6 +55,10 @@ public class DepartmentServiceImpl implements DepartmentService {
         if (!department.getName().equals(departmentDetails.getName()) &&
                 departmentRepository.existsByNameAndCompany(departmentDetails.getName(), employee.getCompany())) {
             throw new IllegalArgumentException("이미 존재하는 이름입니다");
+        }
+
+        if(!UserRole.MANAGER.equals(employee.getRole())){
+            throw new AccessDeniedException("권한이 없습니다");
         }
 
         department.updateName(departmentDetails.getName());
@@ -45,9 +69,19 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public Long deleteDepartment(Long id) {
+    @Caching(
+            evict = {
+                    @CacheEvict(value = "department", key = "#id"),
+                    @CacheEvict(value = "departments", key = "#employee.company.id", allEntries = true)
+            }
+    )
+    public Long deleteDepartment(Long id, Employee employee) {
         Department department = departmentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("부서를 찾을 수 없습니다 " + id));
+
+        if(!UserRole.MANAGER.equals(employee.getRole())){
+            throw new AccessDeniedException("권한이 없습니다");
+        }
 
         departmentRepository.delete(department);
